backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic[email]==2.5.3

# HTTP Requests (for Gemini API)
requests==2.31.0

# Gmail API Integration
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Firebase Admin SDK
firebase-admin==6.4.0

# HTML Parsing (for email body extraction)
beautifulsoup4==4.12.3

# Additional Utilities
python-multipart==0.0.6